% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellchatV2.R
\name{cellchatV2_path_visu}
\alias{cellchatV2_path_visu}
\title{Run CellChatV2 Pathway Visualization}
\usage{
cellchatV2_path_visu(output_dir, species, pathway_to_show)
}
\arguments{
\item{output_dir}{Path to the folder where the visualization results will be saved (figures and tables).}

\item{species}{The species of the data, either 'human' or 'mouse'.}

\item{pathway_to_show}{A specific pathways to visualize.}
}
\description{
This function performs pathway visualization for the CellChatV2 analysis. 
It generates various visualizations, including signaling strength plots, heatmaps, scatter plots, 
LR gene expression violin plots, and communication probability bubble plots for specified pathways. 
Users can define the number of top pathways to visualize and specify pathways of interest that may not 
have been included in the top_n most significant pathways already visualized.
}
