% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_transfer.R
\name{doTransferLabel}
\alias{doTransferLabel}
\title{Transfer cell labels from one Seurat object to another}
\usage{
doTransferLabel(X, Y, varToHarmonize = NULL, transferCoordinates = FALSE)
}
\arguments{
\item{X}{Seurat object used as reference for the transfer of cell labels}

\item{Y}{Seurat object in which the cell labels will be transferred}

\item{varToHarmonize}{a character string specifying a variable to harmonize data over}

\item{transferCoordinates}{Logical. If TRUE, the UMAP coordinates of the query object will be
updated with the ones from the reference object.}
}
\value{
A Seurat object with the transferred cell labels and UMAP coordinates (if transferCoordinates = TRUE)
}
\description{
This function transfers cell labels from a reference Seurat object to a query Seurat object.
The labels are transferred based on k-NN classification of the query cells in the reference UMAP space.
}
