% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo.R
\name{run_scvelo}
\alias{run_scvelo}
\title{Function to run the scVelo pipeline using velociraptor}
\usage{
run_scvelo(
  seurat_obj,
  loom_files = NULL,
  output_dir = ".",
  loom_file_subset_by = NULL,
  loom_file_subset_column = "orig.ident",
  annotation_column = NULL,
  mode = "stochastic",
  grid_resolutions = c(50),
  arrow_sizes = c(0.5, 1),
  vector_widths = c(0.25, 0.5),
  time_point = list(),
  time_point_column = NULL,
  color_scale = NULL,
  name_by = NULL
)
}
\arguments{
\item{seurat_obj}{Seurat object containing the scRNA-seq data (Required)}

\item{loom_files}{path and file names of Spliced and unspliced counts of the scRNA-seq data (Required if 
not provided in `seurat_obj`). For multiple loom files, the file names in `loom_files` must match the metadata 
column in the Seurat object specified by `loom_file_subset_column`}

\item{output_dir}{A character vector specifying the output directory}

\item{loom_file_subset_by}{A character variable specifying how the Seurat object should be subsetted in order 
to match each indivdidual loom file name. This vector must correspond to a metadata column specified by 
`loom_file_subset_column` in the Seurat object, and the order of `loom_file_subset_by` must match the order of 
loom_files. If `loom_file_subset_by` is not provided (default `loom_file_subset_by <- NULL`), it will be 
automatically extracted from file names in `loom_files` to ensure the correct order. 
If there is only one loom file provided, this variable should be left blank: `loom_file_subset_by <- NULL`}

\item{loom_file_subset_column}{A string specifying the name of the metadata column in the Seurat object that
should be used for subsetting to match each of the loom files. This column must exist in the Seurat object 
metadata. If there is only one loom file provided, this #' variable should be left blank: 
`loom_file_subset_column <- NULL`; if there are multiple loom files provided, this variable needs to be provided.}

\item{annotation_column}{A string specifying the metdata column name of cell type annotations. Default: 
`annotation_column <- NULL`, and `Idents(seurat_obj)` is used as the cell type annotation}

\item{mode}{Mode for scVelo velocity calculation, default: stochastic; can be one of four options: "steady_state" 
(original), "deterministic", "stochastic" (fastest: recommended, default), "dynamical"}

\item{grid_resolutions}{A vector of integers specifying the number of grids along each axis, essentially 
controlling the number of vectors on umap, default 50.}

\item{arrow_sizes}{A vector of integers or floats controlling velocity vector size (arrow head size), default 0.5}

\item{vector_widths}{A vector of integers or floats controlling velocity vector size (vector width), default 0.5}

\item{time_point}{A list of character vectors representing a group of time points used to calculate RNA velocity 
together, can be left blank}

\item{time_point_column}{A string specify the name of the metadata column in the Seurat object that should be 
used for subsetting the group of time points}

\item{color_scale}{A character vector of colors to be used in plotting, must match number of unique values in 
the metadata column marked by @name_by}

\item{name_by}{A string specify the name of the metadata column in the Seurat object that should be used for colors}
}
\value{
A list of scVelo data objects
}
\description{
This function performs RNA velocity calculations from .loom files with the scVelo package.
In this function, users can calculate RNA velocity of the whole data as well as a subset of time points.
}
