% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_scvelo_full.R
\name{run_scvelo_full}
\alias{run_scvelo_full}
\title{Function to run the scVelo pipeline using python script}
\usage{
run_scvelo_full(
  h5ad_file = "scvelo/rds/obj_spliced_unspliced.h5ad",
  output_dir = ".",
  annotation_column = "ID",
  mode = "stochastic",
  top_gene = 5
)
}
\arguments{
\item{h5ad_file}{input h5ad file path and name, if running after run_scvelo(), this object has a fixed name and does not need to be changed}

\item{output_dir}{A character vector specifying the output directory}

\item{annotation_column}{A character variable specifying which metadata column of the h5ad object contains cell type annotations}

\item{mode}{Mode to conduct scvelo velocity calculation, either 'stochastic (default)', 'deterministic', or 'dynamical (slowest)'}

\item{top_gene}{The number of top differential velocity genes to plot phase portrait for}
}
\value{
A list of scVelo data objects
}
\description{
This function performs RNA velocity calculations from .h5ad file using scVelo python script.
Workflow of this function: 
1. calculate RNA velocity using scVelo workflow
2. cluster-specific differential velocity genes
3. trajectory inference using PAGA
This function outputs basic figures such as spliced/unspliced count proportion, projected RNA velocity on umap,
the phase portrait (ratio of spliced/unspliced RNA abundance) for top differential genes, 
and directed graphs of predicted lineages from PAGA trajectory inference
}
